+++ article/serializers.py
@@ -1,5 +1,29 @@
 from rest_framework import serializers
-from article.models import Article
+from article.models import Article, Comment
+
+
+class CommentSerializer(serializers.ModelSerializer):
+    """
+    comment/ url에 GET방식일 때 사용합니다.
+    comment db에 저장된 모든 게시글을 보여줍니다.
+    DateTimeField를 사용하여 시간 가독성을 좋게했습니다.
+    """
+
+    created_at = serializers.DateTimeField(format="%m월%d일 %H:%M", read_only=True)
+    author = serializers.SerializerMethodField()
+
+    def get_author(self, obj):
+        return obj.author.username
+
+    class Meta:
+        model = Comment
+        fields = "__all__"
+
+
+class CommentCreateSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Comment
+        fields = ("content",)
 
 
 class ArticleSerializer(serializers.ModelSerializer):
